% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\docType{data}
\name{ref_immgen}
\alias{ref_immgen}
\title{Matrix of average gene expression from immgen data}
\format{
An object of class \code{function} of length 1.
}
\source{
["SingleR"]
}
\usage{
ref_immgen
}
\description{
253 mouse immune cell types
}
\examples{
ref_immgen(metadata = TRUE)
}
\seealso{
Other ref: 
\code{\link{ref_MCA}()},
\code{\link{ref_cortex_dev}()},
\code{\link{ref_hema_microarray}()},
\code{\link{ref_moca_main}()},
\code{\link{ref_mouse.rnaseq}()},
\code{\link{ref_pan_indrop}()},
\code{\link{ref_pan_smartseq2}()},
\code{\link{ref_tabula_muris_drop}()},
\code{\link{ref_tabula_muris_facs}()}
}
\concept{ref}
\keyword{datasets}
