% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\docType{data}
\name{ref_mouse_atlas}
\alias{ref_mouse_atlas}
\title{Atlas matrix of average gene expression aggregated from scRNA-seq data of mouse organisms}
\format{
An object of class \code{function} of length 1.
}
\source{
<https://github.com/rnabioco/scRNA-seq-Cell-Ref-Matrix/blob/master/atlas/musMusculus/MouseAtlas.rds>
}
\usage{
ref_mouse_atlas
}
\description{
Data aggregated from NCBI Gene Expression Omnibus
}
\examples{
ref_mouse_atlas(metadata = True) 
}
\seealso{
Other ref: 
\code{\link{ref_MCA}()},
\code{\link{ref_cortex_dev}()},
\code{\link{ref_hema_microarray}()},
\code{\link{ref_immgen}()},
\code{\link{ref_moca_main}()},
\code{\link{ref_mouse.rnaseq}()},
\code{\link{ref_pan_indrop}()},
\code{\link{ref_pan_smartseq2}()},
\code{\link{ref_tabula_muris_drop}()},
\code{\link{ref_tabula_muris_facs}()}
}
\concept{ref}
\keyword{datasets}
